// src/app/tax-report/tax-report.enum.ts
export enum TaxReportErrorLabel {
  CouldNotReloadTaxReports = 'Oops. Could not reload tax reports. Please try again later.',
  CouldNotDeleteTaxReport = 'Oops. We could not delete tax report. Please try again later.',
  CouldNotCreateTaxReport = 'Oops. We could not create tax report. Please try again later.',
  CouldNotDownloadTaxReport = 'Oops. We could not download your file. Please try disabling your Pop-up blocker and try again.'
}